+++ ArduPlane/quadplane.cpp
@@ -730,7 +730,7 @@ bool QuadPlane::setup(void)
 
     // setup the trim of any motors used by AP_Motors so I/O board
     // failsafe will disable motors
-    uint16_t mask = plane.quadplane.motors->get_motor_mask();
+    uint32_t mask = plane.quadplane.motors->get_motor_mask();
     hal.rcout->set_failsafe_pwm(mask, plane.quadplane.motors->get_pwm_output_min());
 
     // default QAssist state as set with Q_OPTIONS
